make todo done only if there is no pre todo not done yet

+++ src/main/scala/com/jinhyuk/http4sdoobiecircetodo/Todo.scala
@@ -3,7 +3,7 @@ package com.jinhyuk.http4sdoobiecircetodo
 import io.circe.generic.extras.ConfiguredJsonCodec
 import io.circe.generic.extras.Configuration
 
-@ConfiguredJsonCodec case class Todo(id: Int, name: String, isDone: Boolean, isDeleted: Boolean)
+@ConfiguredJsonCodec case class Todo(id: Long, name: String, isDone: Boolean, isDeleted: Boolean)
 
 object Todo {
   implicit private val config: Configuration = Configuration.default.withSnakeCaseMemberNames
